Refactored project to reuse db client in multiple docker services. Also changed startup process
